requirement.txt
aioice==0.10.2
aiortc==1.14.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
Authlib==1.6.9
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
deprecation==2.1.0
dnspython==2.8.0
easydict==1.13
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.1
google-crc32c==1.8.0
grpcio==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.3
insightface==0.7.3
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
onnxruntime==1.24.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
simsimd==6.5.16
six==1.17.0
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
validators==0.35.0
wcwidth==0.6.0
weaviate-client==4.20.4
absl-py==2.4.0
aioice==0.10.2
aiortc==1.14.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
Authlib==1.6.9
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
deprecation==2.1.0
dnspython==2.8.0
easydict==1.13
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.1
google-crc32c==1.8.0
grpcio==1.78.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.3
insightface==0.7.3
jax==0.9.2
jaxlib==0.9.2
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
mediapipe==0.10.14
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
onnxruntime==1.24.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
simsimd==6.5.16
six==1.17.0
sounddevice==0.5.5
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
validators==0.35.0
wcwidth==0.6.0
weaviate-client==4.20.4
absl-py==2.4.0
aioice==0.10.2
aiortc==1.14.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
Authlib==1.6.9
av==16.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
deprecation==2.1.0
dnspython==2.8.0
easydict==1.13
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.1
google-crc32c==1.8.0
grpcio==1.78.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ifaddr==0.2.0
ImageIO==2.37.3
insightface==0.7.3
jax==0.9.2
jaxlib==0.9.2
joblib==1.5.3
kiwisolver==1.5.0
lazy-loader==0.5
matplotlib==3.10.8
mediapipe==0.10.14
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
onnx==1.20.1
onnxruntime==1.24.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
prettytable==3.17.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pylibsrtp==1.0.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
simsimd==6.5.16
six==1.17.0
sounddevice==0.5.5
starlette==0.52.1
stringzilla==4.6.0
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
validators==0.35.0
wcwidth==0.6.0
weaviate-client==4.20.4
